# docs/swagger.yaml
definitions:
  api.countUsersResponse:
    properties:
      count:
        type: integer
    type: object
  api.createUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        description: |-
          The Role of the user
          example: 1 for admin, 2 for user
        type: integer
      surname:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - role
    - surname
    - username
    type: object
  api.loginUserRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.loginUserResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/api.userResponse'
    type: object
  api.userResponse:
    properties:
      created_at:
        type: string
      email:
        description: |-
          The Email of the user
          example: joe.doe@email.com
        type: string
      enabled:
        description: |-
          The enabled/disabeld flag
          example: 0 for disabled, 1 for enabled
        type: boolean
      name:
        description: |-
          The Name of he user
          example: Some name
        type: string
      password_change_at:
        type: string
      role:
        description: |-
          The Role of the user
          example: 1 for admin, 2 for user
        type: integer
      surname:
        description: |-
          The Surname of he user
          example: Some name
        type: string
      username:
        description: |-
          The username of a thing
          example: joedoe
        type: string
    type: object
  api.verifyEmailRequest:
    properties:
      email_id:
        type: integer
      secret_code:
        type: string
    required:
    - email_id
    - secret_code
    type: object
  api.verifyEmailResponse:
    properties:
      is_verified:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /admin/users:
    get:
      consumes:
      - application/json
      description: get paginated user list
      parameters:
      - description: Authorization
        in: header
        name: authorization
        required: true
        type: string
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.userResponse'
            type: array
      summary: get the user list paginated
      tags:
      - users
  /admin/users/count:
    get:
      description: get user count
      parameters:
      - description: Authorization
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.countUsersResponse'
      summary: get the user count for pagination
      tags:
      - users
  /user/{username}:
    get:
      description: returns the user info
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Authorization
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
      summary: get the user info
      tags:
      - users
  /users:
    post:
      consumes:
      - application/json
      description: create a new user
      parameters:
      - description: createUserRequest
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
      summary: create a new user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: returns a new PASETO token and the logged user info
      parameters:
      - description: loginUserRequest
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.loginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginUserResponse'
        "404":
          description: no rows in resultset
      summary: perform a new user login
      tags:
      - users
  /users/logout:
    post:
      description: delete the user session
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
      summary: perform a user logout
      tags:
      - users
  /verify_email:
    get:
      consumes:
      - application/json
      description: check the code received in the email during registration is correct
      parameters:
      - description: verifyEmailRequest
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.verifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.verifyEmailResponse'
        "500":
          description: failed to verify email
      summary: perform the user verification with email check
      tags:
      - users
swagger: "2.0"